+++ lib/easy_meli/errors.rb
@@ -9,4 +9,10 @@ module EasyMeli
   end
 
   class AuthenticationError < Error; end
+  
+  class TooManyRequestsError < Error
+    def initialize(response)
+      super('Too many requests', response)
+    end
+  end
 end