+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # Changelog
 
 ## V 0.5.0
-- Renamed all `refresh_token` methods to access Token to be consistent
-- `self.api_client` Prevent access_token override when initialize togehter with a refresh_token
+- `self.api_client` Prevent access_token override when initialize together with a refresh_token.
+- Raises a EasyMeli::TooManyRequestsError if a 429 response status is returned.